--- .firecrest-demo-config.json
@@ -1,8 +1,0 @@
-{
-    "url": "http://localhost:8000/",
-    "token_uri": "http://localhost:8080/auth/realms/kcrealm/protocol/openid-connect/token",
-    "client_id": "firecrest-sample",
-    "client_secret": "b391e177-fa50-4987-beaf-e6d33ca93571",
-    "machine": "cluster",
-    "scratch_path": "/home/service-account-firecrest-sample"
-}
