+++ .github/workflows/tests.yml
@@ -49,7 +49,7 @@ jobs:
         python -m pip install --upgrade pip
         pip install -e .[dev]
     - name: Test with pytest
-      run: pytest -vv --firecrest-requests --cov=aiida_firecrest --cov-report=xml --cov-report=term
+      run: pytest -vv  --cov=aiida_firecrest --cov-report=xml --cov-report=term
 
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
