+++ pyproject.toml
@@ -26,7 +26,7 @@ requires-python = ">=3.9"
 dependencies = [
     "aiida-core@git+https://github.com/khsrali/aiida-core.git@aiida-firecrest#egg=aiida-core",
     "click",
-    "pyfirecrest~=2.5.0",
+    "pyfirecrest@git+https://github.com/khsrali/pyfirecrest.git@main#egg=pyfirecrest",
     "pyyaml",
     "requests",
 ]
