+++ tests/conftest.py
@@ -1,142 +1,369 @@
-"""Pytest configuration for the aiida-firecrest tests.
-
-This sets up the Firecrest server to use, and telemetry for API requests.
-"""
-from __future__ import annotations
-
-from functools import partial
-from json import load as json_load
+import hashlib
+import os
 from pathlib import Path
-from typing import Any, Callable
-from urllib.parse import urlparse
+import random
+import stat
+from typing import Optional
 
-from _pytest.terminal import TerminalReporter
-import firecrest as f7t
+from aiida import orm
+import firecrest
+import firecrest.path
 import pytest
-import requests
-import yaml
-
-from aiida_firecrest.utils_test import FirecrestConfig, FirecrestMockServer
 
 
-def pytest_report_header(config):
-    if config.getoption("--firecrest-config"):
-        header = [
-            "Running against FirecREST server: {}".format(
-                config.getoption("--firecrest-config")
-            )
-        ]
-        if config.getoption("--firecrest-no-clean"):
-            header.append("Not cleaning up FirecREST server after tests!")
-        return header
-    return ["Running against Mock FirecREST server"]
+class Values:
+    _DEFAULT_PAGE_SIZE: int = 25
 
 
-def pytest_terminal_summary(
-    terminalreporter: TerminalReporter, exitstatus: int, config: pytest.Config
-):
-    """Called after all tests have run."""
-    data = config.stash.get("firecrest_requests", None)
-    if data is None:
-        return
-    terminalreporter.write(
-        yaml.dump(
-            {"Firecrest requests telemetry": data},
-            default_flow_style=False,
-            sort_keys=True,
-        )
+@pytest.fixture(name="firecrest_computer")
+def _firecrest_computer(myfirecrest, tmpdir: Path):
+    """Create and return a computer configured for Firecrest.
+
+    Note, the computer is not stored in the database.
+    """
+
+    # create a temp directory and set it as the workdir
+    _scratch = tmpdir / "scratch"
+    _temp_directory = tmpdir / "temp"
+    _scratch.mkdir()
+    _temp_directory.mkdir()
+
+    Path(tmpdir / ".firecrest").mkdir()
+    _secret_path = Path(tmpdir / ".firecrest/secret69")
+    _secret_path.write_text("SECRET_STRING")
+
+    computer = orm.Computer(
+        label="test_computer",
+        description="test computer",
+        hostname="-",
+        workdir=str(_scratch),
+        transport_type="firecrest",
+        scheduler_type="firecrest",
     )
+    computer.set_minimum_job_poll_interval(5)
+    computer.set_default_mpiprocs_per_machine(1)
+    computer.configure(
+        url=" https://URI",
+        token_uri="https://TOKEN_URI",
+        client_id="CLIENT_ID",
+        client_secret=str(_secret_path),
+        client_machine="MACHINE_NAME",
+        small_file_size_mb=1.0,
+        temp_directory=str(_temp_directory),
+    )
+    return computer
+
+
+class MockFirecrest:
+    def __init__(self, firecrest_url, *args, **kwargs):
+        self._firecrest_url = firecrest_url
+        self.args = args
+        self.kwargs = kwargs
+
+        self.whoami = whomai
+        self.list_files = list_files
+        self.stat = stat_
+        self.mkdir = mkdir
+        self.simple_delete = simple_delete
+        self.parameters = parameters
+        self.symlink = symlink
+        self.checksum = checksum
+        self.simple_download = simple_download
+        self.simple_upload = simple_upload
+        self.compress = compress
+        self.extract = extract
+        self.copy = copy
+        self.submit = submit
+        self.poll_active = poll_active
+
+
+class MockClientCredentialsAuth:
+    def __init__(self, *args, **kwargs):
+        self.args = args
+        self.kwargs = kwargs
 
 
 @pytest.fixture(scope="function")
-def firecrest_server(
+def myfirecrest(
     pytestconfig: pytest.Config,
-    request: pytest.FixtureRequest,
     monkeypatch,
-    tmp_path: Path,
 ):
-    """A fixture which provides a mock Firecrest server to test against."""
-    config_path: str | None = request.config.getoption("--firecrest-config")
-    no_clean: bool = request.config.getoption("--firecrest-no-clean")
-    record_requests: bool = request.config.getoption("--firecrest-requests")
-    telemetry: RequestTelemetry | None = None
+    monkeypatch.setattr(firecrest, "Firecrest", MockFirecrest)
+    monkeypatch.setattr(firecrest, "ClientCredentialsAuth", MockClientCredentialsAuth)
 
-    if config_path is not None:
-        # if given, use this config
-        with open(config_path, encoding="utf8") as handle:
-            config = json_load(handle)
-        config = FirecrestConfig(**config)
-        # rather than use the scratch_path directly, we use a subfolder,
-        # which we can then clean
-        config.scratch_path = config.scratch_path + "/pytest_tmp"
 
-        # we need to connect to the client here,
-        # to ensure that the scratch path exists and is empty
-        client = f7t.Firecrest(
-            firecrest_url=config.url,
-            authorization=f7t.ClientCredentialsAuth(
-                config.client_id, config.client_secret, config.token_uri
-            ),
+def submit(
+    machine: str,
+    script_str: Optional[str] = None,
+    script_remote_path: Optional[str] = None,
+    script_local_path: Optional[str] = None,
+    local_file=False,
+):
+    if local_file:
+        raise DeprecationWarning("local_file is not supported")
+
+    if script_remote_path and not Path(script_remote_path).exists():
+        raise FileNotFoundError(f"File {script_remote_path} does not exist")
+    job_id = random.randint(10000, 99999)
+    return {"jobid": job_id}
+
+
+def poll_active(machine: str, jobs: list[str], page_number: int = 0):
+    response = []
+    # 12 satets are defined in firecrest
+    states = [
+        "TIMEOUT",
+        "SUSPENDED",
+        "PREEMPTED",
+        "CANCELLED",
+        "NODE_FAIL",
+        "PENDING",
+        "FAILED",
+        "RUNNING",
+        "CONFIGURING",
+        "QUEUED",
+        "COMPLETED",
+        "COMPLETING",
+    ]
+    for i in range(len(jobs)):
+        response.append(
+            {
+                "job_data_err": "",
+                "job_data_out": "",
+                "job_file": "somefile.sh",
+                "job_file_err": "somefile-stderr.txt",
+                "job_file_out": "somefile-stdout.txt",
+                "job_info_extra": "Job info returned successfully",
+                "jobid": f"{jobs[i]}",
+                "name": "aiida-45",
+                "nodelist": "nid00049",
+                "nodes": "1",
+                "partition": "normal",
+                "start_time": "0:03",
+                "state": states[i % 12],
+                "time": "2024-06-21T10:44:42",
+                "time_left": "29:57",
+                "user": "Prof. Wang",
+            }
         )
-        client.mkdir(config.machine, config.scratch_path, p=True)
 
-        if record_requests:
-            telemetry = RequestTelemetry()
-            monkeypatch.setattr(requests, "get", partial(telemetry.wrap, requests.get))
-            monkeypatch.setattr(
-                requests, "post", partial(telemetry.wrap, requests.post)
+    return response[
+        page_number
+        * Values._DEFAULT_PAGE_SIZE : (page_number + 1)
+        * Values._DEFAULT_PAGE_SIZE
+    ]
+
+
+def whomai(machine: str):
+    assert machine == "MACHINE_NAME"
+    return "test_user"
+
+
+def list_files(
+    machine: str, target_path: str, recursive: bool = False, show_hidden: bool = False
+):
+    # this is mimiking the expected behaviour from the firecrest code.
+
+    content_list = []
+    for root, dirs, files in os.walk(target_path):
+        if not recursive and root != target_path:
+            continue
+        for name in dirs + files:
+            full_path = os.path.join(root, name)
+            relative_path = (
+                Path(os.path.relpath(root, target_path)).joinpath(name).as_posix()
             )
-            monkeypatch.setattr(requests, "put", partial(telemetry.wrap, requests.put))
-            monkeypatch.setattr(
-                requests, "delete", partial(telemetry.wrap, requests.delete)
+            if os.path.islink(full_path):
+                content_type = "l"
+                link_target = (
+                    os.readlink(full_path) if os.path.islink(full_path) else None
+                )
+            elif os.path.isfile(full_path):
+                content_type = "-"
+                link_target = None
+            elif os.path.isdir(full_path):
+                content_type = "d"
+                link_target = None
+            else:
+                content_type = "NON"
+                link_target = None
+            permissions = stat.filemode(Path(full_path).lstat().st_mode)[1:]
+            if name.startswith(".") and not show_hidden:
+                continue
+            content_list.append(
+                {
+                    "name": relative_path,
+                    "type": content_type,
+                    "link_target": link_target,
+                    "permissions": permissions,
+                }
             )
 
-        yield config
-        # Note this shouldn't really work, for folders but it does :shrug:
-        # because they use `rm -r`:
-        # https://github.com/eth-cscs/firecrest/blob/7f02d11b224e4faee7f4a3b35211acb9c1cc2c6a/src/utilities/utilities.py#L347
-        if not no_clean:
-            client.simple_delete(config.machine, config.scratch_path)
+    return content_list
+
+
+def stat_(machine: str, targetpath: firecrest.path, dereference=True):
+    stats = os.stat(
+        targetpath, follow_symlinks=bool(dereference) if dereference else False
+    )
+    return {
+        "ino": stats.st_ino,
+        "dev": stats.st_dev,
+        "nlink": stats.st_nlink,
+        "uid": stats.st_uid,
+        "gid": stats.st_gid,
+        "size": stats.st_size,
+        "atime": stats.st_atime,
+        "mtime": stats.st_mtime,
+        "ctime": stats.st_ctime,
+    }
+
+
+def mkdir(machine: str, target_path: str, p: bool = False):
+    if p:
+        os.makedirs(target_path)
     else:
-        # otherwise use mock server
-        server = FirecrestMockServer(tmp_path)
-        if record_requests:
-            telemetry = RequestTelemetry()
-            mock_request = partial(telemetry.wrap, server.mock_request)
-        else:
-            mock_request = server.mock_request
-        monkeypatch.setattr(requests, "get", mock_request)
-        monkeypatch.setattr(requests, "post", mock_request)
-        monkeypatch.setattr(requests, "put", mock_request)
-        monkeypatch.setattr(requests, "delete", mock_request)
-        monkeypatch.setattr(requests.Session, "get", mock_request)
-        monkeypatch.setattr(requests.Session, "post", mock_request)
-        monkeypatch.setattr(requests.Session, "put", mock_request)
-        monkeypatch.setattr(requests.Session, "delete", mock_request)
-        yield server.config
-
-    # save data on the server
-    if telemetry is not None:
-        test_name = request.node.name
-        pytestconfig.stash.setdefault("firecrest_requests", {})[
-            test_name
-        ] = telemetry.counts
+        os.mkdir(target_path)
 
 
-class RequestTelemetry:
-    """A to gather telemetry on requests."""
+def simple_delete(machine: str, target_path: str):
+    if not Path(target_path).exists():
+        raise FileNotFoundError(f"File or folder {target_path} does not exist")
+    if os.path.isdir(target_path):
+        os.rmdir(target_path)
+    else:
+        os.remove(target_path)
 
-    def __init__(self) -> None:
-        self.counts = {}
 
-    def wrap(
-        self,
-        method: Callable[..., requests.Response],
-        url: str | bytes,
-        **kwargs: Any,
-    ) -> requests.Response:
-        """Wrap a requests method to gather telemetry."""
-        endpoint = urlparse(url if isinstance(url, str) else url.decode("utf-8")).path
-        self.counts.setdefault(endpoint, 0)
-        self.counts[endpoint] += 1
-        return method(url, **kwargs)
+def symlink(machine: str, target_path: str, link_path: str):
+    # this is how firecrest does it
+    os.system(f"ln -s {target_path}  {link_path}")
+
+
+def simple_download(machine: str, remote_path: str, local_path: str):
+    # this procedure is complecated in firecrest, but I am simplifying it here
+    # we don't care about the details of the download, we just want to make sure
+    # that the aiida-firecrest code is calling the right functions at right time
+    if Path(remote_path).is_dir():
+        raise IsADirectoryError(f"{remote_path} is a directory")
+    if not Path(remote_path).exists():
+        raise FileNotFoundError(f"{remote_path} does not exist")
+    os.system(f"cp {remote_path} {local_path}")
+
+
+def simple_upload(
+    machine: str, local_path: str, remote_path: str, file_name: Optional[str] = None
+):
+    # this procedure is complecated in firecrest, but I am simplifying it here
+    # we don't care about the details of the upload, we just want to make sure
+    # that the aiida-firecrest code is calling the right functions at right time
+    if Path(local_path).is_dir():
+        raise IsADirectoryError(f"{local_path} is a directory")
+    if not Path(local_path).exists():
+        raise FileNotFoundError(f"{local_path} does not exist")
+    if file_name:
+        remote_path = os.path.join(remote_path, file_name)
+    os.system(f"cp {local_path} {remote_path}")
+
+
+def copy(machine: str, source_path: str, target_path: str):
+    # this is how firecrest does it
+    # https://github.com/eth-cscs/firecrest/blob/db6ba4ba273c11a79ecbe940872f19d5cb19ac5e/src/utilities/utilities.py#L451C1-L452C1
+    os.system(f"cp --force -dR --preserve=all -- '{source_path}' '{target_path}'")
+
+
+def compress(
+    machine: str, source_path: str, target_path: str, dereference: bool = True
+):
+    # this is how firecrest does it
+    # https://github.com/eth-cscs/firecrest/blob/db6ba4ba273c11a79ecbe940872f19d5cb19ac5e/src/utilities/utilities.py#L460
+    basedir = os.path.dirname(source_path)
+    file_path = os.path.basename(source_path)
+    deref = "--dereference" if dereference else ""
+    os.system(f"tar {deref} -czf '{target_path}' -C '{basedir}' '{file_path}'")
+
+
+def extract(machine: str, source_path: str, target_path: str):
+    # this is how firecrest does it
+    # https://github.com/eth-cscs/firecrest/blob/db6ba4ba273c11a79ecbe940872f19d5cb19ac5e/src/common/cscs_api_common.py#L1110C18-L1110C65
+    os.system(f"tar -xf '{source_path}' -C '{target_path}'")
+
+
+def checksum(machine: str, remote_path: str) -> int:
+    if not remote_path.exists():
+        return False
+    # Firecrest uses sha256
+    sha256_hash = hashlib.sha256()
+    with open(remote_path, "rb") as f:
+        for byte_block in iter(lambda: f.read(4096), b""):
+            sha256_hash.update(byte_block)
+
+    return sha256_hash.hexdigest()
+
+
+def parameters():
+    # note: I took this from https://firecrest-tds.cscs.ch/ or https://firecrest.cscs.ch/
+    # if code is not working but test passes, it means you need to update this dictionary
+    # with the latest FirecREST parameters
+    return {
+        "compute": [
+            {
+                "description": "Type of resource and workload manager used in compute microservice",
+                "name": "WORKLOAD_MANAGER",
+                "unit": "",
+                "value": "Slurm",
+            }
+        ],
+        "storage": [
+            {
+                "description": "Type of object storage, like `swift`, `s3v2` or `s3v4`.",
+                "name": "OBJECT_STORAGE",
+                "unit": "",
+                "value": "s3v4",
+            },
+            {
+                "description": "Expiration time for temp URLs.",
+                "name": "STORAGE_TEMPURL_EXP_TIME",
+                "unit": "seconds",
+                "value": "86400",
+            },
+            {
+                "description": "Maximum file size for temp URLs.",
+                "name": "STORAGE_MAX_FILE_SIZE",
+                "unit": "MB",
+                "value": "5120",
+            },
+            {
+                "description": "Available filesystems through the API.",
+                "name": "FILESYSTEMS",
+                "unit": "",
+                "value": [
+                    {
+                        "mounted": ["/project", "/store", "/scratch/snx3000tds"],
+                        "system": "dom",
+                    },
+                    {
+                        "mounted": ["/project", "/store", "/capstor/scratch/cscs"],
+                        "system": "pilatus",
+                    },
+                ],
+            },
+        ],
+        "utilities": [
+            {
+                "description": "The maximum allowable file size for various operations of the utilities microservice",
+                "name": "UTILITIES_MAX_FILE_SIZE",
+                "unit": "MB",
+                "value": "69",
+            },
+            {
+                "description": (
+                    "Maximum time duration for executing the commands "
+                    "in the cluster for the utilities microservice."
+                ),
+                "name": "UTILITIES_TIMEOUT",
+                "unit": "seconds",
+                "value": "5",
+            },
+        ],
+    }
