--- tests/test_scheduler/test_write_script_full.txt
@@ -1,20 +1,0 @@
-#!/bin/bash
-#SBATCH -H
-#SBATCH --requeue
-#SBATCH --mail-user=True
-#SBATCH --mail-type=BEGIN
-#SBATCH --mail-type=FAIL
-#SBATCH --mail-type=END
-#SBATCH --job-name="test_job"
-#SBATCH --get-user-env
-#SBATCH --output=test.out
-#SBATCH --error=test.err
-#SBATCH --partition=test_queue
-#SBATCH --account=test_account
-#SBATCH --qos=test_qos
-#SBATCH --nice=100
-#SBATCH --nodes=1
-#SBATCH --ntasks-per-node=1
-#SBATCH --time=01:00:00
-#SBATCH --mem=1
-test_command
