--- tests/test_scheduler/test_write_script_minimal.txt
@@ -1,5 +1,0 @@
-#!/bin/bash
-#SBATCH --no-requeue
-#SBATCH --error=slurm-%j.err
-#SBATCH --nodes=1
-#SBATCH --ntasks-per-node=1
