+++ tests/test_transport.py
@@ -1,153 +1,627 @@
-"""Tests isolating only the Transport."""
+import os
 from pathlib import Path
-import platform
+from unittest.mock import patch
 
+from aiida import orm
 import pytest
 
-from aiida_firecrest.transport import FirecrestTransport
-from aiida_firecrest.utils_test import FirecrestConfig
+
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_mkdir(firecrest_computer: orm.Computer, tmpdir: Path):
+    transport = firecrest_computer.get_transport()
+
+    _scratch = tmpdir / "sampledir"
+    transport.mkdir(_scratch)
+    assert _scratch.exists()
+
+    _scratch = tmpdir / "sampledir2" / "subdir"
+    transport.makedirs(_scratch)
+    assert _scratch.exists()
 
 
-@pytest.fixture(name="transport")
-def _transport(firecrest_server: FirecrestConfig):
-    transport = FirecrestTransport(
-        url=firecrest_server.url,
-        token_uri=firecrest_server.token_uri,
-        client_id=firecrest_server.client_id,
-        client_secret=firecrest_server.client_secret,
-        client_machine=firecrest_server.machine,
-        small_file_size_mb=firecrest_server.small_file_size_mb,
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_is_file(firecrest_computer: orm.Computer, tmpdir: Path):
+    transport = firecrest_computer.get_transport()
+
+    _scratch = tmpdir / "samplefile"
+    Path(_scratch).touch()
+    assert transport.isfile(_scratch)
+    assert not transport.isfile(_scratch / "does_not_exist")
+
+
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_is_dir(firecrest_computer: orm.Computer, tmpdir: Path):
+    transport = firecrest_computer.get_transport()
+
+    _scratch = tmpdir / "sampledir"
+    _scratch.mkdir()
+
+    assert transport.isdir(_scratch)
+    assert not transport.isdir(_scratch / "does_not_exist")
+
+
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_normalize(firecrest_computer: orm.Computer):
+    transport = firecrest_computer.get_transport()
+    assert transport.normalize("/path/to/dir") == os.path.normpath("/path/to/dir")
+    assert transport.normalize("path/to/dir") == os.path.normpath("path/to/dir")
+    assert transport.normalize("path/to/dir/") == os.path.normpath("path/to/dir/")
+    assert transport.normalize("path/to/../dir") == os.path.normpath("path/to/../dir")
+    assert transport.normalize("path/to/../../dir") == os.path.normpath(
+        "path/to/../../dir"
     )
-    transport.chdir(firecrest_server.scratch_path)
-    yield transport
+    assert transport.normalize("path/to/../../dir/") == os.path.normpath(
+        "path/to/../../dir/"
+    )
+    assert transport.normalize("path/to/../../dir/../") == os.path.normpath(
+        "path/to/../../dir/../"
+    )
 
 
-def test_path_exists(firecrest_server: FirecrestConfig, transport: FirecrestTransport):
-    assert transport.path_exists(firecrest_server.scratch_path)
-    assert not transport.path_exists(firecrest_server.scratch_path + "/file.txt")
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_remove(firecrest_computer: orm.Computer, tmpdir: Path):
+    transport = firecrest_computer.get_transport()
+
+    _scratch = tmpdir / "samplefile"
+    Path(_scratch).touch()
+    transport.remove(_scratch)
+    assert not _scratch.exists()
+
+    _scratch = tmpdir / "sampledir"
+    _scratch.mkdir()
+    transport.rmtree(_scratch)
+    assert not _scratch.exists()
+
+    _scratch = tmpdir / "sampledir"
+    _scratch.mkdir()
+    Path(_scratch / "samplefile").touch()
+    with pytest.raises(OSError):
+        transport.rmdir(_scratch)
+
+    os.remove(_scratch / "samplefile")
+    transport.rmdir(_scratch)
+    assert not _scratch.exists()
 
 
-def test_get_attribute(
-    firecrest_server: FirecrestConfig, transport: FirecrestTransport
-):
-    transport._cwd.joinpath("test.txt").touch()
-    attrs = transport.get_attribute(firecrest_server.scratch_path + "/test.txt")
-    assert set(attrs) == {
-        "st_size",
-        "st_atime",
-        "st_mode",
-        "st_gid",
-        "st_mtime",
-        "st_uid",
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_symlink(firecrest_computer: orm.Computer, tmpdir: Path):
+    transport = firecrest_computer.get_transport()
+
+    _scratch = Path(tmpdir / "samplefile-2sym")
+    Path(_scratch).touch()
+    _symlink = Path(tmpdir / "samplelink")
+    transport.symlink(_scratch, _symlink)
+    assert _symlink.is_symlink()
+    assert _symlink.resolve() == _scratch
+
+
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_listdir(firecrest_computer: orm.Computer, tmpdir: Path):
+    transport = firecrest_computer.get_transport()
+
+    _scratch = tmpdir / "sampledir"
+    _scratch.mkdir()
+    # to test basics
+    Path(_scratch / "file1").touch()
+    Path(_scratch / "dir1").mkdir()
+    Path(_scratch / ".hidden").touch()
+    # to test recursive
+    Path(_scratch / "dir1" / "file2").touch()
+
+    assert set(transport.listdir(_scratch)) == {"file1", "dir1", ".hidden"}
+    assert set(transport.listdir(_scratch, recursive=True)) == {
+        "file1",
+        "dir1",
+        ".hidden",
+        "dir1/file2",
     }
-    assert isinstance(attrs.st_mode, int)
+    # to test symlink
+    Path(_scratch / "dir1" / "dir2").mkdir()
+    Path(_scratch / "dir1" / "dir2" / "file3").touch()
+    os.symlink(_scratch / "dir1" / "dir2", _scratch / "dir2_link")
+    os.symlink(_scratch / "dir1" / "file2", _scratch / "file_link")
+
+    assert set(transport.listdir(_scratch, recursive=True)) == {
+        "file1",
+        "dir1",
+        ".hidden",
+        "dir1/file2",
+        "dir1/dir2",
+        "dir1/dir2/file3",
+        "dir2_link",
+        "file_link",
+    }
+
+    assert set(transport.listdir(_scratch / "dir2_link", recursive=False)) == {"file3"}
 
 
-def test_isdir(firecrest_server: FirecrestConfig, transport: FirecrestTransport):
-    assert transport.isdir(firecrest_server.scratch_path)
-    assert not transport.isdir(firecrest_server.scratch_path + "/other")
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_get(firecrest_computer: orm.Computer, tmpdir: Path):
+    """
+    This is minimal test is to check if get() is raising errors as expected,
+    and directing to getfile() and gettree() as expected.
+    Mainly just checking error handeling and folder creation.
+    """
+    transport = firecrest_computer.get_transport()
+
+    _remote = tmpdir / "remotedir"
+    _local = tmpdir / "localdir"
+    _remote.mkdir()
+    _local.mkdir()
+
+    # check if the code is directing to getfile() or gettree() as expected
+    with patch.object(transport, "gettree", autospec=True) as mock_gettree:
+        transport.get(_remote, _local)
+    mock_gettree.assert_called_once()
+
+    with patch.object(transport, "gettree", autospec=True) as mock_gettree:
+        os.symlink(_remote, tmpdir / "dir_link")
+        transport.get(tmpdir / "dir_link", _local)
+    mock_gettree.assert_called_once()
+
+    with patch.object(transport, "getfile", autospec=True) as mock_getfile:
+        Path(_remote / "file1").write_text("file1")
+        transport.get(_remote / "file1", _local / "file1")
+    mock_getfile.assert_called_once()
+
+    with patch.object(transport, "getfile", autospec=True) as mock_getfile:
+        os.symlink(_remote / "file1", _remote / "file1_link")
+        transport.get(_remote / "file1_link", _local / "file1_link")
+    mock_getfile.assert_called_once()
+
+    # raise if remote file/folder does not exist
+    with pytest.raises(FileNotFoundError):
+        transport.get(_remote / "does_not_exist", _local)
+    transport.get(_remote / "does_not_exist", _local, ignore_nonexisting=True)
+
+    # raise if localpath is relative
+    with pytest.raises(ValueError):
+        transport.get(_remote, Path(_local).relative_to(tmpdir))
+    with pytest.raises(ValueError):
+        transport.get(_remote / "file1", Path(_local).relative_to(tmpdir))
 
 
-def test_mkdir(firecrest_server: FirecrestConfig, transport: FirecrestTransport):
-    transport.mkdir(firecrest_server.scratch_path + "/test")
-    assert transport.isdir(firecrest_server.scratch_path + "/test")
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_getfile(firecrest_computer: orm.Computer, tmpdir: Path):
+    transport = firecrest_computer.get_transport()
+
+    _remote = tmpdir / "remotedir"
+    _local = tmpdir / "localdir"
+    _remote.mkdir()
+    _local.mkdir()
+
+    Path(_remote / "file1").write_text("file1")
+    Path(_remote / ".hidden").write_text(".hidden")
+    os.symlink(_remote / "file1", _remote / "file1_link")
+
+    # raise if remote file does not exist
+    with pytest.raises(FileNotFoundError):
+        transport.getfile(_remote / "does_not_exist", _local)
+
+    # raise if localfilename not provided
+    with pytest.raises(IsADirectoryError):
+        transport.getfile(_remote / "file1", _local)
+
+    # raise if localpath is relative
+    with pytest.raises(ValueError):
+        transport.getfile(_remote / "file1", Path(_local / "file1").relative_to(tmpdir))
+
+    # don't mix up directory with file
+    with pytest.raises(FileNotFoundError):
+        transport.getfile(_remote, _local / "file1")
+
+    # write where I tell you to
+    transport.getfile(_remote / "file1", _local / "file1")
+    transport.getfile(_remote / "file1", _local / "file1-prime")
+    assert Path(_local / "file1").read_text() == "file1"
+    assert Path(_local / "file1-prime").read_text() == "file1"
+
+    # always overwrite
+    transport.getfile(_remote / "file1", _local / "file1")
+    assert Path(_local / "file1").read_text() == "file1"
+
+    Path(_local / "file1").write_text("notfile1")
+
+    transport.getfile(_remote / "file1", _local / "file1")
+    assert Path(_local / "file1").read_text() == "file1"
+
+    # don't skip hidden files
+    transport.getfile(_remote / ".hidden", _local / ".hidden-prime")
+    assert Path(_local / ".hidden-prime").read_text() == ".hidden"
+
+    # follow links
+    transport.getfile(_remote / "file1_link", _local / "file1_link")
+    assert Path(_local / "file1_link").read_text() == "file1"
+    assert not Path(_local / "file1_link").is_symlink()
 
 
-def test_large_file_transfers(
-    firecrest_server: FirecrestConfig, transport: FirecrestTransport, tmp_path: Path
-):
-    """Large file transfers (> 5MB by default) have to be downloaded/uploaded via a different pathway."""
-    content = "a" * (transport._small_file_size_bytes + 1)
+@pytest.mark.parametrize("payoff", [True, False])
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_gettree(firecrest_computer: orm.Computer, tmpdir: Path, payoff: bool):
+    """
+    This test is to check `gettree` through non tar mode.
+    bytar= True in this test.
+    """
+    transport = firecrest_computer.get_transport()
+    transport.payoff_override = payoff
 
-    # upload
-    remote_path = firecrest_server.scratch_path + "/file.txt"
-    assert not transport.isfile(remote_path)
-    file_path = tmp_path.joinpath("file.txt")
-    file_path.write_text(content)
-    transport.putfile(str(file_path), remote_path)
-    assert transport.isfile(remote_path)
+    # Note:
+    # SSH transport behaviour, 69 is a directory
+    # transport.get('somepath/69', 'someremotepath/') == transport.get('somepath/69', 'someremotepath')
+    # transport.get('somepath/69', 'someremotepath/') == transport.get('somepath/69/', 'someremotepath/')
+    # transport.get('someremotepath/69', 'somepath/69')--> if 69 exist, create 69 inside it ('somepath/69/69')
+    # transport.get('someremotepath/69', 'somepath/69')--> if 69 no texist,create 69 inside it ('somepath/69')
+    # transport.get('somepath/69', 'someremotepath/6889/69') --> create everything, make_parent = True
 
-    # download
-    if transport._url.startswith("http://localhost") and platform.system() == "Darwin":
-        pytest.skip("Skipping large file download test on macOS with localhost server.")
-        # TODO this is a known issue whereby a 403 is returned when trying to download the supplied file url
-        # due to a signature mismatch
-    new_path = tmp_path.joinpath("file2.txt")
-    assert not new_path.is_file()
-    transport.getfile(remote_path, new_path)
-    assert new_path.is_file()
-    assert new_path.read_text() == content
+    _remote = tmpdir / "remotedir"
+    _local = tmpdir / "localdir"
+    _remote.mkdir()
+    _local.mkdir()
+    # a typical tree
+    Path(_remote / "file1").write_text("file1")
+    Path(_remote / ".hidden").write_text(".hidden")
+    Path(_remote / "dir1").mkdir()
+    Path(_remote / "dir1" / "file2").write_text("file2")
+    # with symlinks
+    Path(_remote / "dir2").mkdir()
+    Path(_remote / "dir2" / "file3").write_text("file3")
+    os.symlink(_remote / "file1", _remote / "dir1" / "file1_link")
+    os.symlink(_remote / "dir2", _remote / "dir1" / "dir2_link")
+    # if symlinks are pointing to a relative path
+    os.symlink(Path("../file1"), _remote / "dir1" / "file10_link")
+    os.symlink(Path("../dir2"), _remote / "dir1" / "dir20_link")
 
+    # raise if remote file does not exist
+    with pytest.raises(OSError):
+        transport.gettree(_remote / "does_not_exist", _local)
 
-def test_listdir(firecrest_server: FirecrestConfig, transport: FirecrestTransport):
-    assert transport.listdir(firecrest_server.scratch_path) == []
+    # raise if local is a file
+    Path(tmpdir / "isfile").touch()
+    with pytest.raises(OSError):
+        transport.gettree(_remote, tmpdir / "isfile")
 
+    # raise if localpath is relative
+    with pytest.raises(ValueError):
+        transport.gettree(_remote, Path(_local).relative_to(tmpdir))
 
-def test_copyfile(firecrest_server: FirecrestConfig, transport: FirecrestTransport):
-    from_path = firecrest_server.scratch_path + "/copy_from.txt"
-    to_path = firecrest_server.scratch_path + "/copy_to.txt"
+    # If destination directory does not exists, AiiDA expects transport make the new path as root not _remote.name
+    transport.gettree(_remote, _local / "newdir")
+    _root = _local / "newdir"
+    # tree should be copied recursively
+    assert Path(_root / "file1").read_text() == "file1"
+    assert Path(_root / ".hidden").read_text() == ".hidden"
+    assert Path(_root / "dir1" / "file2").read_text() == "file2"
+    assert Path(_root / "dir2" / "file3").read_text() == "file3"
+    # symlink should be followed
+    assert Path(_root / "dir1" / "file1_link").read_text() == "file1"
+    assert Path(_root / "dir1" / "dir2_link" / "file3").read_text() == "file3"
+    assert Path(_root / "dir1" / "file10_link").read_text() == "file1"
+    assert Path(_root / "dir1" / "dir20_link" / "file3").read_text() == "file3"
+    assert not Path(_root / "dir1" / "file1_link").is_symlink()
+    assert not Path(_root / "dir1" / "dir2_link" / "file3").is_symlink()
+    assert not Path(_root / "dir1" / "file10_link").is_symlink()
+    assert not Path(_root / "dir1" / "dir20_link" / "file3").is_symlink()
 
-    transport.write_binary(from_path, b"test")
-
-    assert not transport.path_exists(to_path)
-    transport.copyfile(from_path, to_path)
-    assert transport.isfile(to_path)
-    assert transport.read_binary(to_path) == b"test"
-
-
-def test_copyfile_symlink_noderef(
-    firecrest_server: FirecrestConfig, transport: FirecrestTransport
-):
-    from_path = firecrest_server.scratch_path + "/copy_from.txt"
-    from_path_symlink = firecrest_server.scratch_path + "/copy_from_symlink.txt"
-    to_path = firecrest_server.scratch_path + "/copy_to_symlink.txt"
-
-    transport.write_binary(from_path, b"test")
-    transport.symlink(from_path, from_path_symlink)
-
-    assert not transport.path_exists(to_path)
-    transport.copyfile(from_path_symlink, to_path, dereference=False)
-    assert transport.isfile(to_path)
-    assert transport.read_binary(to_path) == b"test"
-
-
-def test_copyfile_symlink_deref(
-    firecrest_server: FirecrestConfig, transport: FirecrestTransport
-):
-    from_path = firecrest_server.scratch_path + "/copy_from.txt"
-    from_path_symlink = firecrest_server.scratch_path + "/copy_from_symlink.txt"
-    to_path = firecrest_server.scratch_path + "/copy_to_symlink.txt"
-
-    transport.write_binary(from_path, b"test")
-    transport.symlink(from_path, from_path_symlink)
-
-    assert not transport.path_exists(to_path)
-    transport.copyfile(from_path_symlink, to_path, dereference=True)
-    assert transport.isfile(to_path)
-    assert transport.read_binary(to_path) == b"test"
-
-
-def test_remove(firecrest_server: FirecrestConfig, transport: FirecrestTransport):
-    transport.write_binary(firecrest_server.scratch_path + "/file.txt", b"test")
-    assert transport.path_exists(firecrest_server.scratch_path + "/file.txt")
-    transport.remove(firecrest_server.scratch_path + "/file.txt")
-    assert not transport.path_exists(firecrest_server.scratch_path + "/file.txt")
-
-
-def test_rmtree(firecrest_server: FirecrestConfig, transport: FirecrestTransport):
-    transport.mkdir(firecrest_server.scratch_path + "/test")
-    transport.write_binary(firecrest_server.scratch_path + "/test/file.txt", b"test")
-    assert transport.path_exists(firecrest_server.scratch_path + "/test/file.txt")
-    transport.rmtree(firecrest_server.scratch_path + "/test")
-    assert not transport.path_exists(firecrest_server.scratch_path + "/test")
+    # If destination directory does exists, AiiDA expects transport make _remote.name and write into it
+    # however this might have changed in the newer versions of AiiDA ~ 2.6.0 (IDK)
+    transport.gettree(_remote, _local / "newdir")
+    _root = _local / "newdir" / Path(_remote).name
+    # tree should be copied recursively
+    assert Path(_root / "file1").read_text() == "file1"
+    assert Path(_root / ".hidden").read_text() == ".hidden"
+    assert Path(_root / "dir1" / "file2").read_text() == "file2"
+    assert Path(_root / "dir2" / "file3").read_text() == "file3"
+    # symlink should be followed
+    assert Path(_root / "dir1" / "file1_link").read_text() == "file1"
+    assert Path(_root / "dir1" / "dir2_link" / "file3").read_text() == "file3"
+    assert Path(_root / "dir1" / "file10_link").read_text() == "file1"
+    assert Path(_root / "dir1" / "dir20_link" / "file3").read_text() == "file3"
+    assert not Path(_root / "dir1" / "file1_link").is_symlink()
+    assert not Path(_root / "dir1" / "dir2_link" / "file3").is_symlink()
+    assert not Path(_root / "dir1" / "file10_link").is_symlink()
+    assert not Path(_root / "dir1" / "dir20_link" / "file3").is_symlink()
 
 
-def test_rename(firecrest_server: FirecrestConfig, transport: FirecrestTransport):
-    transport.write_binary(firecrest_server.scratch_path + "/file.txt", b"test")
-    assert transport.path_exists(firecrest_server.scratch_path + "/file.txt")
-    transport.rename(
-        firecrest_server.scratch_path + "/file.txt",
-        firecrest_server.scratch_path + "/file2.txt",
-    )
-    assert not transport.path_exists(firecrest_server.scratch_path + "/file.txt")
-    assert transport.path_exists(firecrest_server.scratch_path + "/file2.txt")
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_put(firecrest_computer: orm.Computer, tmpdir: Path):
+    """
+    This is minimal test is to check if put() is raising errors as expected,
+    and directing to putfile() and puttree() as expected.
+    Mainly just checking error handeling and folder creation.
+    """
+    transport = firecrest_computer.get_transport()
+
+    _remote = tmpdir / "remotedir"
+    _local = tmpdir / "localdir"
+    _remote.mkdir()
+    _local.mkdir()
+
+    # check if the code is directing to putfile() or puttree() as expected
+    with patch.object(transport, "puttree", autospec=True) as mock_puttree:
+        transport.put(_local, _remote)
+    mock_puttree.assert_called_once()
+
+    with patch.object(transport, "puttree", autospec=True) as mock_puttree:
+        os.symlink(_local, tmpdir / "dir_link")
+        transport.put(tmpdir / "dir_link", _remote)
+    mock_puttree.assert_called_once()
+
+    with patch.object(transport, "putfile", autospec=True) as mock_putfile:
+        Path(_local / "file1").write_text("file1")
+        transport.put(_local / "file1", _remote / "file1")
+    mock_putfile.assert_called_once()
+
+    with patch.object(transport, "putfile", autospec=True) as mock_putfile:
+        os.symlink(_local / "file1", _local / "file1_link")
+        transport.put(_local / "file1_link", _remote / "file1_link")
+    mock_putfile.assert_called_once()
+
+    # raise if local file/folder does not exist
+    with pytest.raises(FileNotFoundError):
+        transport.put(_local / "does_not_exist", _remote)
+    transport.put(_local / "does_not_exist", _remote, ignore_nonexisting=True)
+
+    # raise if localpath is relative
+    with pytest.raises(ValueError):
+        transport.put(Path(_local).relative_to(tmpdir), _remote)
+    with pytest.raises(ValueError):
+        transport.put(Path(_local / "file1").relative_to(tmpdir), _remote)
+
+
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_putfile(firecrest_computer: orm.Computer, tmpdir: Path):
+    transport = firecrest_computer.get_transport()
+
+    _remote = tmpdir / "remotedir"
+    _local = tmpdir / "localdir"
+    _remote.mkdir()
+    _local.mkdir()
+
+    Path(_local / "file1").write_text("file1")
+    Path(_local / ".hidden").write_text(".hidden")
+    os.symlink(_local / "file1", _local / "file1_link")
+
+    # raise if local file does not exist
+    with pytest.raises(FileNotFoundError):
+        transport.putfile(_local / "does_not_exist", _remote)
+
+    # raise if remotefilename is not provided
+    with pytest.raises(ValueError):
+        transport.putfile(_local / "file1", _remote)
+
+    # raise if localpath is relative
+    with pytest.raises(ValueError):
+        transport.putfile(Path(_local / "file1").relative_to(tmpdir), _remote / "file1")
+
+    # don't mix up directory with file
+    with pytest.raises(ValueError):
+        transport.putfile(_local, _remote / "file1")
+
+    # write where I tell you to
+    transport.putfile(_local / "file1", _remote / "file1")
+    transport.putfile(_local / "file1", _remote / "file1-prime")
+    assert Path(_remote / "file1").read_text() == "file1"
+    assert Path(_remote / "file1-prime").read_text() == "file1"
+
+    # always overwrite
+    transport.putfile(_local / "file1", _remote / "file1")
+    assert Path(_remote / "file1").read_text() == "file1"
+
+    Path(_remote / "file1").write_text("notfile1")
+
+    transport.putfile(_local / "file1", _remote / "file1")
+    assert Path(_remote / "file1").read_text() == "file1"
+
+    # don't skip hidden files
+    transport.putfile(_local / ".hidden", _remote / ".hidden-prime")
+    assert Path(_remote / ".hidden-prime").read_text() == ".hidden"
+
+    # follow links
+    transport.putfile(_local / "file1_link", _remote / "file1_link")
+    assert Path(_remote / "file1_link").read_text() == "file1"
+    assert not Path(_remote / "file1_link").is_symlink()
+
+
+@pytest.mark.parametrize("payoff", [True, False])
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_puttree(firecrest_computer: orm.Computer, tmpdir: Path, payoff: bool):
+    """
+    This test is to check `puttree` through non tar mode.
+    payoff= False in this test, so just checking if putting files one by one is working as expected.
+    """
+    transport = firecrest_computer.get_transport()
+    transport.payoff_override = payoff
+
+    # Note:
+    # SSH transport behaviour
+    # transport.put('somepath/69', 'someremotepath/') == transport.put('somepath/69', 'someremotepath')
+    # transport.put('somepath/69', 'someremotepath/') != transport.put('somepath/69/', 'someremotepath/')
+    # transport.put('somepath/69', 'someremotepath/67') --> if 67 not exist, create and move content 69
+    #  inside it (someremotepath/67)
+    # transport.put('somepath/69', 'someremotepath/67') --> if 67 exist, create 69 inside it (someremotepath/67/69)
+    # transport.put('somepath/69', 'someremotepath/6889/69')  -->  useless Error: OSError
+    # Weired
+    # SSH bug:
+    # transport.put('somepath/69', 'someremotepath/') --> assuming someremotepath exists, make 69
+    # while
+    # transport.put('somepath/69/', 'someremotepath/') --> assuming someremotepath exists, OSError:
+    # cannot make someremotepath
+
+    _remote = tmpdir / "remotedir"
+    _local = tmpdir / "localdir"
+    _remote.mkdir()
+    _local.mkdir()
+    # a typical tree
+    Path(_local / "dir1").mkdir()
+    Path(_local / "dir2").mkdir()
+    Path(_local / "file1").write_text("file1")
+    Path(_local / ".hidden").write_text(".hidden")
+    Path(_local / "dir1" / "file2").write_text("file2")
+    Path(_local / "dir2" / "file3").write_text("file3")
+    # with symlinks to a file even if pointing to a relative path
+    os.symlink(_local / "file1", _local / "dir1" / "file1_link")
+    os.symlink(Path("../file1"), _local / "dir1" / "file10_link")
+    # with symlinks to a folder even if pointing to a relative path
+    os.symlink(_local / "dir2", _local / "dir1" / "dir2_link")
+    os.symlink(Path("../dir2"), _local / "dir1" / "dir20_link")
+
+    # raise if local file does not exist
+    with pytest.raises(OSError):
+        transport.puttree(_local / "does_not_exist", _remote)
+
+    # raise if local is a file
+    with pytest.raises(ValueError):
+        Path(tmpdir / "isfile").touch()
+        transport.puttree(tmpdir / "isfile", _remote)
+
+    # raise if localpath is relative
+    with pytest.raises(ValueError):
+        transport.puttree(Path(_local).relative_to(tmpdir), _remote)
+
+    # If destination directory does not exists, AiiDA expects transport make the new path as root not using _local.name
+    transport.puttree(_local, _remote / "newdir")
+    _root = _remote / "newdir"
+    # tree should be copied recursively
+    assert Path(_root / "file1").read_text() == "file1"
+    assert Path(_root / ".hidden").read_text() == ".hidden"
+    assert Path(_root / "dir1" / "file2").read_text() == "file2"
+    assert Path(_root / "dir2" / "file3").read_text() == "file3"
+    # symlink should be followed
+    assert Path(_root / "dir1" / "file1_link").read_text() == "file1"
+    assert Path(_root / "dir1" / "dir2_link" / "file3").read_text() == "file3"
+    assert Path(_root / "dir1" / "file10_link").read_text() == "file1"
+    assert Path(_root / "dir1" / "dir20_link" / "file3").read_text() == "file3"
+    assert not Path(_root / "dir1" / "file1_link").is_symlink()
+    assert not Path(_root / "dir1" / "dir2_link" / "file3").is_symlink()
+    assert not Path(_root / "dir1" / "file10_link").is_symlink()
+    assert not Path(_root / "dir1" / "dir20_link" / "file3").is_symlink()
+
+    # If destination directory does exists, AiiDA expects transport make _local.name and write into it
+    # however this might have changed in the newer versions of AiiDA ~ 2.6.0 (IDK)
+    transport.puttree(_local, _remote / "newdir")
+    _root = _remote / "newdir" / Path(_local).name
+    # tree should be copied recursively
+    assert Path(_root / "file1").read_text() == "file1"
+    assert Path(_root / ".hidden").read_text() == ".hidden"
+    assert Path(_root / "dir1" / "file2").read_text() == "file2"
+    assert Path(_root / "dir2" / "file3").read_text() == "file3"
+    # symlink should be followed
+    assert Path(_root / "dir1" / "file1_link").read_text() == "file1"
+    assert Path(_root / "dir1" / "dir2_link" / "file3").read_text() == "file3"
+    assert Path(_root / "dir1" / "file10_link").read_text() == "file1"
+    assert Path(_root / "dir1" / "dir20_link" / "file3").read_text() == "file3"
+    assert not Path(_root / "dir1" / "file1_link").is_symlink()
+    assert not Path(_root / "dir1" / "dir2_link" / "file3").is_symlink()
+    assert not Path(_root / "dir1" / "file10_link").is_symlink()
+    assert not Path(_root / "dir1" / "dir20_link" / "file3").is_symlink()
+
+
+@pytest.mark.parametrize("to_test", ["copy", "copytree"])
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_copy(firecrest_computer: orm.Computer, tmpdir: Path, to_test: str):
+    transport = firecrest_computer.get_transport()
+    if to_test == "copy":
+        testing = transport.copy
+    elif to_test == "copytree":
+        testing = transport.copytree
+
+    _remote_1 = tmpdir / "remotedir_1"
+    _remote_2 = tmpdir / "remotedir_2"
+    _remote_1.mkdir()
+    _remote_2.mkdir()
+
+    # raise if source or destination does not exist
+    with pytest.raises(FileNotFoundError):
+        testing(_remote_1 / "does_not_exist", _remote_2)
+    with pytest.raises(FileNotFoundError):
+        testing(_remote_1, _remote_2 / "does_not_exist")
+
+    # raise if source is inappropriate
+    if to_test == "copytree":
+        Path(tmpdir / "file1").touch()
+        with pytest.raises(ValueError):
+            testing(tmpdir / "file1", _remote_2)
+
+    # a typical tree
+    Path(_remote_1 / "dir1").mkdir()
+    Path(_remote_1 / "dir2").mkdir()
+    Path(_remote_1 / "file1").write_text("file1")
+    Path(_remote_1 / ".hidden").write_text(".hidden")
+    Path(_remote_1 / "dir1" / "file2").write_text("file2")
+    Path(_remote_1 / "dir2" / "file3").write_text("file3")
+    # with symlinks to a file even if pointing to a relative path
+    os.symlink(_remote_1 / "file1", _remote_1 / "dir1" / "file1_link")
+    os.symlink(Path("../file1"), _remote_1 / "dir1" / "file10_link")
+    # with symlinks to a folder even if pointing to a relative path
+    os.symlink(_remote_1 / "dir2", _remote_1 / "dir1" / "dir2_link")
+    os.symlink(Path("../dir2"), _remote_1 / "dir1" / "dir20_link")
+
+    testing(_remote_1, _remote_2)
+
+    _root_2 = _remote_2 / Path(_remote_1).name
+    # tree should be copied recursively
+    assert Path(_root_2 / "dir1").exists()
+    assert Path(_root_2 / "dir2").exists()
+    assert Path(_root_2 / "file1").read_text() == "file1"
+    assert Path(_root_2 / ".hidden").read_text() == ".hidden"
+    assert Path(_root_2 / "dir1" / "file2").read_text() == "file2"
+    assert Path(_root_2 / "dir2" / "file3").read_text() == "file3"
+    # symlink should be followed
+    assert Path(_root_2 / "dir1" / "dir2_link").exists()
+    assert Path(_root_2 / "dir1" / "dir20_link").exists()
+    assert Path(_root_2 / "dir1" / "file1_link").read_text() == "file1"
+    assert Path(_root_2 / "dir1" / "file10_link").read_text() == "file1"
+    assert Path(_root_2 / "dir1" / "file1_link").is_symlink()
+    assert Path(_root_2 / "dir1" / "dir2_link").is_symlink()
+    assert Path(_root_2 / "dir1" / "file10_link").is_symlink()
+    assert Path(_root_2 / "dir1" / "dir20_link").is_symlink()
+
+
+@pytest.mark.usefixtures("aiida_profile_clean")
+def test_copyfile(firecrest_computer: orm.Computer, tmpdir: Path):
+    transport = firecrest_computer.get_transport()
+    testing = transport.copyfile
+
+    _remote_1 = tmpdir / "remotedir_1"
+    _remote_2 = tmpdir / "remotedir_2"
+    _remote_1.mkdir()
+    _remote_2.mkdir()
+
+    # raise if source or destination does not exist
+    with pytest.raises(FileNotFoundError):
+        testing(_remote_1 / "does_not_exist", _remote_2)
+    # in this case don't raise and just create the file
+    Path(tmpdir / "_").touch()
+    testing(tmpdir / "_", _remote_2 / "does_not_exist")
+
+    # raise if source is unappropriate
+    with pytest.raises(ValueError):
+        testing(tmpdir, _remote_2)
+
+    # a typical tree
+    Path(_remote_1 / "file1").write_text("file1")
+    Path(_remote_1 / ".hidden").write_text(".hidden")
+    # with symlinks to a file even if pointing to a relative path
+    os.symlink(_remote_1 / "file1", _remote_1 / "file1_link")
+    os.symlink(Path("file1"), _remote_1 / "file10_link")
+
+    # write where I tell you to
+    testing(_remote_1 / "file1", _remote_2 / "file1")
+    assert Path(_remote_2 / "file1").read_text() == "file1"
+
+    # always overwrite
+    Path(_remote_2 / "file1").write_text("notfile1")
+    testing(_remote_1 / "file1", _remote_2 / "file1")
+    assert Path(_remote_2 / "file1").read_text() == "file1"
+
+    # don't skip hidden files
+    testing(_remote_1 / ".hidden", _remote_2 / ".hidden-prime")
+    assert Path(_remote_2 / ".hidden-prime").read_text() == ".hidden"
+
+    # preserve links and don't follow them
+    testing(_remote_1 / "file1_link", _remote_2 / "file1_link")
+    assert Path(_remote_2 / "file1_link").read_text() == "file1"
+    assert Path(_remote_2 / "file1_link").is_symlink()
+    testing(_remote_1 / "file10_link", _remote_2 / "file10_link")
+    assert Path(_remote_2 / "file10_link").read_text() == "file1"
+    assert Path(_remote_2 / "file10_link").is_symlink()
